GetToThePoint: Gossip/quest auto completing mod
Skips quests that you add, currently includes all of the Alterac Valley quests by default you can add things to skip by holding ALT down and clicking the text you want to skip. If the quest has item requirements you'll need to go to the completion window (The part where it lists the items required/amount) before it'll start auto skipping it.

Automagically completes the quest with the highest item requirements, even if you add it yourself. So if you have 55 Sunfury Signets, it'll turn them in using the 10 turn in quest 5 times, then use the single turn in quest the remaining 5.